Working on a port of DEV regularization to Theano.
The DEV regularization method is now working in Theano, with the
underlying network and dropout implementation based heavily on code by
Misha Desnil, which was itself based on code from the deeplearning.net
tutorials. I wanted a sanity check on the performance of my method, and
a greater degree of certainty that the underlying network
implementation was more in line with what other people are currently
doing.

Test performance seems good, as compared to standard dropout, and the
Theano-based implementation seems noticeably faster on the GPU than my
gnumpy code. Theano takes a bit of getting used to though.